refactor(Toolbar): add Typescript types
